refactor - 라인 폴더로 이동, Stretegy 호출되게 변경, of 메소드 생성, IntStream을 통한 라인 생성으로 변경
